Refactoring the XMLRPC::Client error-handling
I split it all into smaller, manageable chunks,
and used methods for each step, instead
of having one huge call.

Note that I made all of the tests first, then
refactored the code, so I'm confident there's no
behavior change.

I don't know that this is actually a lot cleaner,
but it seems that way to me.  I'm open to
skipping this, but I think it makes the whole thing
a lot cleaner.

Signed-off-by: Luke Kanies <luke@madstop.com>