Using the FileCollection where appropriate.
This commit just replaces the :file and :line accessors
with the use of the new FileCollection Lookup module.
This should mean that we've normalized all file names in
a given process, which *might* have drastic RAM improvements.

For initial simplicity, I've gone with a single global
collection of file names, but it's built so it's easy to use
individual file collections instead.

Signed-off-by: Luke Kanies <luke@madstop.com>