Refactoring how the Settings file is parsed
The goal of this refactor was to use a cached attribute
for the LoadedFile instance we use to monitor whether
the file needs reparsing.  We were getting tests that
affected later tests because they were holding on to
LoadedFile stubs, somehow.

The other main change here is that the Settings#parse
method now knows how to look up its own file path.

Signed-off-by: Luke Kanies <luke@madstop.com>