Fixing every failing test I can find on the build server.
All but one of these tests is fixed by:
* Stubbing Puppet.settings.use
* Adding /usr/sbin to PATH

The only other one was the package integration test,
which stupidly assumed a default was specified in the test.
The fix here is twofold: Remove that assumption (the test is
now 'pending' if no default is available), and add a default
for Ubuntu.  (The default is in the test, not the default
package provider - that is, it's testing the default package
provider, thus it can't rely on that information.)

Signed-off-by: Luke Kanies <luke@madstop.com>