Passing checksums around instead of file contents
This switches the file's 'content' parameter to always
use checksums, rather than always using content but switching
to checksums whenever necessary.

This greatly simplifies all the logging requirements (so
that content doesn't show up in logs), but also simplifies
insync comparisons, and much more.

In the process, I found that the code was pulling down file content
more often than was necessary, and fixing that cut 40% off of the time
of a very small transaction.

Signed-off-by: Luke Kanies <luke@madstop.com>