Fixing #1949 - relationships now use attributes instead of a label
This was important because the use of the label to store attributes
was a holdover from the GRATR library, and if we didn't cease its
use before we switched to RESTful catalogs, then we'd be stuck with
the @label instance variable forever, essentially.

Now we can add and remove variables however we please.

Signed-off-by: Luke Kanies <luke@madstop.com>