Fixing #1885 - Relationships metaparams do not cascade
Because we now pass catalogs around, rather than a tree
of resources, we no longer lose the metaparam information
in definitions and classes.  Thus, we no longer need
to pass them down to contained resources.

Ideally we'd remove cascading of all metaparams (which is
ticket #1903) but 'schedule' and 'noop' are inherently
recursive but not in a way that the graph support can currently
easily solve, so that's going to have to wait for a later
release.

Signed-off-by: Luke Kanies <luke@madstop.com>