Allow REST auth system to restrict an ACL to authenticated or unauthenticated request
Introduces a new auth.conf directive (auth or authenticated) which
takes an argument (on,yes/off,no/all,any).
This can be used to restrict an ACL to only some state of
authentication of a REST request, or any.
If no auth directive is given, the ACL will only trigger for
authenticated requests.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>