Fixing #2187 - Puppet::Resource is expected by Rails support
We previously used and expected Puppet::Parser::Resource
instances, but 0.25 converts them all to Puppet::Resource
instances before they're passed out of the compiler,
so the Rails integration had to be changed to expect that.

There's still some muddling, because the rails resources
only generate parser resources, but that works for now
because that's what we expect when collecting resources.

Signed-off-by: Luke Kanies <luke@madstop.com>