Adding a Spec lib directory and moving tmpfile to it
We had a common pattern for creating a temporary
file during integration tests, and this just
makes that common pattern explicit by
moving it to a module in the newly-created
lib directory in the spec directory.

We definitely don't want to go overboard in
using libraries in our tests, but sometimes
it gets a bit excessive to completely avoid them.

Signed-off-by: Luke Kanies <luke@madstop.com>