Fix #2220 - Make sure stat is refreshed while managing Files
Since the introduction of the attribute cacher, Files couldn't refresh
their 'stat' attribute while applying the properties.
This could led to misunderstanding (like not setting mode) because
the stat didn't reflect the physical state of the entity, especially
when creating directories, where puppet was taking decision thinking
the directory didn't exist although it was just created.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>