Make sure virtual and rails query use tags when tag are searched
Up to now, when trying to match with tags:
File<<| tag == 'value' |>>
in fact we were querying parameters. Hopefully all the user tags
are stored in parameters so it was working.
But it wasn't possible to search on auto-tags (like class name).
This patch makes sure searching by tag is done on tags both on the
rails side and the resource side.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>