Make sure the parser sees the correct line number
Careful inspection of the parser code show that when we
associate a source line number for an AST node, we use the
current line number of the currently lexed token.
In many case, this is correct, but there are some cases where
this is incorrect.
Unfortunately due to how LALR parser works the ast node creation
of a statement can appear _after_ we lexed another token after
the current statement:

1. $foo = 1
2.
3. class test

When the parser asks for the class token, it can reduce the
assignement statement into the AST VarDef node, because no other
grammar rule match. Unfortunately we already lexed the class token
so we affect to the VarDef node the line number 3 instead of 1.

This is not a real issue for error reporting, but becomes a real
concern when we associate documentation comments to AST node for
puppetdoc.

The solution is to enhance the tokens lexed and returned to the parser
to carry their declaration line number.
Thus a token value becomes a hash: { :value => tokenvalue, :line }

Next, each time we create an AST node, we use the line number of
the correct token (ie the foo line number in the previous example).

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>