Feature #2378 - Implement "thin_storeconfigs"
Thin storeconfigs is a limited version of storeconfigs that is
more performant and still allows the exported/collected resources
system wich is the primary use of storeconfigs.

It works by storing to the database only the exported resources, tags
and host facts.

Since usually those exported resources are less than the number
of total resources for a node, it is expected to be faster than
regular storeconfigs (especially for the first run).

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>