Fix #2392 - use Content-Type for REST communication
There were two problems:
 * server->client communications is using Content-Type with the
direct format name instead of the format mime-type.
 * client->server communications is not using Content-Type to
send the format of the serialized object. Instead it is using the
first member of the Accept header. The Accept header is usually
reserved for the other side, ie what the client will accept
when the server will respond.

This patch makes sure s->c communication contains correct Content-Type
headers.

This patch also adds a Content-Type header containing the mime-type of
the object sent by the client when saving.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>