Fixes #2367 - Mongrel::HTTPRequest returns a StringIO object
When the PUT body is large enough that Mongrel::HTTPRequest#body returns a StringIO object instead of a String. StringIO#to_s then returns "<StringIO#8236987299>" instead of the string contents.

When that string is passed to YAML it returns false which is then passed to save_object without any real time checking.

This is a combination of patches from Jordan Curzon and Ricky Zhou.