Fix #2391 - Exported resources never make to the storeconfigs db
The issue is that when we convert Puppet::Parser::Resource catalog
to a Puppet::Resource catalog before storing it to the database,
we don't allow virtual resource to be converted.
Unfortunately exported resources are virtual by design, and as
such aren't converted, and we lose them, so it isn't possible
to store them in the database.
Unfortunately, the client will get the exported resources too.

The fix is dual-fold:
 * we make sure exported resource are skipped when the transaction is
applied as a last safeguard
 * we filter-out the catalog through the catalog compiler terminus before
the catalog is returned to the client

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>