Fix #2424 - File server can't find module in environment
Actually, the issue is:
*  when the web server gets the request, it creates an indirection
request, filling attributes like ip or node from the HTTP request.
To do this, all the interesting attributes are given in a hash
(called options, see P::I::Request#new).
Once the request is properly initialized the options hash doesn't
contain the ip or node information (see set_attributes)

* the request is then transmitted to the file_serving layer,
which happily wants to use the node attribute to find environments or
perform authorization.
Unfortunately it fetches the node value from the request options hash,
not the request itself.
Since this node information is empty, puppet fails to find the
proper mount point, and fails the download.

This change makes sure we pass all the way down the node and fix
the authorization check.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>