Fixing #2423 - no more strange dependency cycles
We were getting strange dependency cycles because our class structure
mirrored our scope structure.  We can't change the scope structure
without switching from a dynamically scoped language to a lexically scoped
language, which is too big of a change to make right now.  Instead,
I'm changing the resource graph so that all classes default to just
having an edge to the 'main' graph.

This will be a behaviour change for many, in that you were getting
automatic dependencies from this old behaviour, but it will bring
consistency.

Signed-off-by: Luke Kanies <luke@madstop.com>