Minimal patch to fix #2290 (/tmp permissions)
This patch rectifies the specified problem by resetting the
permissions to 1777 after the offending tests have run.  I
am providing it for testing but suspect there may be a deeper
bug hiding behind this one--specifically, it isn't clear to
me that an analogous problem could not occur in production use
(which this patch to the spec would not fix).

This patch also fixes the specific test isolation failure which was
causing ticket #2290; if spec/unit/util/storage.rb was run by
"rake spec" before spec/integration/network/server/webrick.rb
without any intervening tests which did a Puppet.settings.clear,
the the webrick test would change the permissions on /tmp.

Signed-off-by: Markus Roberts <Markus@reality.com>