Fixes #2483 - Log only copies metadata from RAL objects
We were previously trying to figure out what data
was available based on what methods existed, but
that caught a different method profile from
modules.

This fixes it so we only look for this data from
Puppet::Type or Puppet::Parameter instances.

I had to add the ability to skip data that's
not available, since File's 'ensure' parameter
doesn't have 'file' data, I assume because of
the metaprogramming we do around the 'file' value
for 'ensure'.  It's a workaround for now, and there's
a test in there to verify it, anyway.

Signed-off-by: Luke Kanies <luke@madstop.com>