conf/redhat/*.init: Fix condrestart/try-restart
Previously, the Red Hat init scripts used the $pidfile or $lockfile as a
test for whether to restart the daemons.  This caused condrestart to
start the daemons even when they were not running, in cases where they
had died or been killed without cleaning up the $pidfile/$lockfile.

This was reported by Ingvar Hagelund in Red Hat bug #480600.

Signed-off-by: Todd Zullinger <tmz@pobox.com>