Fixes #2503
Added a random suffex to the temporary file name and verify that
the name is not in use (retry on collison).  No additional
tests added as the lifetime of the temporary file is bound by
the routine, making spec driven tests impraticable.

Includes logic to account for dead-end symlinks as well, as per Luke.

Signed-off-by: Markus Roberts <Markus@reality.com>