Fix #2516 - Fix format detection when content-type contains charset
Even though Puppet never transmist a charset information in its
response/request content-type, some proxy (especially Apache with the
infamous AddDefaultCharset configuration) may add this "incorrect"
information.

This patch makes sure that only the mime-type is used when looking
for the format associated with a response or a request.

The patch also provides a better error message when the client or server
code is fed with a request whose mime-type can not be mapped to a known
format.

It also fixes a typo noticed by the original reporter.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>