Fix #2517 - Stack overflow when CA cert missing
This was a "There's A Hole In The Bucket" problem caused when trying
to establish a connection to get a certificate before there was
a certificate with which to establish the connection, ad
infinitum.  The solution was to test for the presence of the
CA cert as well as the host cert before attempting to use them.

This patch modifies existing tests to pass with the new code (by
stubbing out the additional FileTests) and adds a new test which
catches the original problem.

Signed-off-by: Markus Roberts <Markus@reality.com>