Fixing #1544 - plugins in modules now works again
We had to fix the fileserving plumbing to use the request
environment instead of trying to use the node environment.
This was apparently never fixed after we added the environment
to the URI in REST calls.

There's still a bit of refactoring left to clean up the APIs used
in some of this code.

Signed-off-by: Luke Kanies <luke@madstop.com>