Switching the owner/group settings to use symbolic values
We previously allowed the owner and group to be set to
arbitrary values but we never actually used it -- we always
just set them to '$user' or '$group'.  This commit changes
the model to allow 'root' or 'service', where 'service'
is converted to the actual service user/group.

This has the potential to have backward compatibility concerns,
because users could have changed the owner/group in puppet.conf,
but the chances of that are fantastically small.

Signed-off-by: Luke Kanies <luke@madstop.com>