Fixes #2550 Handles case where metadata is nil
The checksum function assumed that if the source was present the
metadata (including checksum) was also, and thus that the type of
the checksum could be extracated with a simple RegExp match.  When
metadata is nil it returns nil for the source checksum and the
modified function passes this on.

Alternatives would be to default to :md5 or to pass on to checking
the :checksum parameter as if there were no source.

Signed-off-by: Markus Roberts <Markus@reality.com>