Fixing #2460 - puppetmasterd can now read the cert and key
This follows on to the previous commits which made Puppet
resilient to the service user being absent.  This just
changes the default owner of all of the cert-related files
to now be the service user, which means that on the server
all of these files will be owned by that user and on
the client, at least when the service user does not exist,
they will be owned by root.

Signed-off-by: Luke Kanies <luke@madstop.com>