Fixing #2590 - modulepath is not cached inappropriately
It was getting cached by Environment instances too early, thus causing some
executables to use the default even when it was overridden.

We're taking the slightly extreme step of clearing the environment
list if any parameter is set, but this is relatively inexpensive and
is the only way to always be correct.

The reason that the environments cache this value at all is that to build
up their module path they have to search through every known module for a lib
or plugins directory, which is then done every time a class is sought in
the language.  Caching it saves a *ton* of file accesses.

Signed-off-by: Luke Kanies <luke@madstop.com>