Fix for #2605 by falling back to alternative solution to #1963
Two solutions were proposed and tested for #1963; both worked but one
(the read_nonblock solution) was used for performance reasons.  This
solution does not work on older ruby implementations (1.8.1) because
read_nonblock is not available.  This patch implements the alternative
fix (IO.popen) as a fallback to handles such cases.

Signed-off-by: Markus Roberts <Markus@reality.com>