Fixs #2620 authconf interpolation, #2570 0-9 in domain names
Partial refactoring to clean up the case in the ticket (host
name containing dots, begining and ending with a digit, was
mistaken for an IP address) and a range of related edge cases.

Stopped short of a full refactoring (put off to 0.26 as #2623)

Added tests for numerous edge cases.

This also fixes the issue raised in #2570.

Signed-off-by: Markus Roberts <Markus@reality.com>