Fix for #2654 (error generating error message)
Since required_features can (and frequently does) return a single
item instead of an Array, the error message needed to be more robust.

The tests were not specific enough to catch the fact that an error
was being raised in the generation of the error, so a more specific
test was added and the required_features accessor test was beefed up
a little.

Signed-off-by: Markus Roberts <Markus@reality.com>

Signed-off-by: Markus Roberts <Markus@reality.com>