Fix #1934 - detailed-exitcodes for puppetd
This option only works when --onetime is specified, as it doesn't make
much sense to worry about exit codes in the context of a long-running
daemon.

This required a refactoring of the existing --detailed-exitcodes code,
as "puppetd" wasn't directly creating a transaction object (like
"puppet" does).

Added Report::exit_status, which did what was previously hard-coded
into the "puppet" executable.

An Agent's "run" method now returns a value (the result of the
individual client class' "run" method)

The "puppetd" agent's "run" method now returns a transaction report, as
that seems like the logical thing to return as the result of applying a
catalog.

Signed-off-by: Deepak Giridharagopal <deepak@brownman.org>