Revised partial fix for #2661 and related issues
If setup code for a process depends on network connectivity
it needs to be protected with a rescue clause as much as the
main body of the process.

Further, Timeout exceptions aren't under StandardError and thus
aren't caught by an un-typed rescue clause.  This doesn't matter
if we've morphed the exception, but will cause the program to
fail if we haven't.

There are many places where these concerns _might_ cause a problem
but in most cases they never will in practice; this patch addesses
the two cases where I have been able to confirm that it actually
can cause the client daemon to exit and two more where I suspect
(but can not prove) that it could.

I'd be willing to push this patch as it stands, as it at least
fixes demonstrable problems.  A more general solution would be
nice.